moved `delete_records` tests to proper file
